move method to user service
